+++ grelmicro/http/_tmf.py
@@ -43,7 +43,7 @@ class TMFError(BaseModel):
     async def charge() -> Charge: ...
     ```
 
-    Read more in the [Problem Details](../http/problems.md) docs.
+    Read more in the [Error Responses](../http/errors.md) docs.
     """
 
     model_config = ConfigDict(extra="allow", populate_by_name=True)